+++ SoA/GasGiantComponentRenderer.cpp
@@ -87,8 +87,8 @@ void GasGiantComponentRenderer::draw(const GasGiantComponent& ggCmp,
     glUniform1f(m_program.getUniform("unInnerRadius"), aCmp->planetRadius);
     glUniform1f(m_program.getUniform("unKrESun"), aCmp->kr * aCmp->esun);
     glUniform1f(m_program.getUniform("unKmESun"), aCmp->km * aCmp->esun);
-    glUniform1f(m_program.getUniform("unKr4PI"), aCmp->kr * M_4_PI);
-    glUniform1f(m_program.getUniform("unKm4PI"), aCmp->km * M_4_PI);
+    glUniform1f(m_program.getUniform("unKr4PI"), (f32)(aCmp->kr * M_4_PI));
+    glUniform1f(m_program.getUniform("unKm4PI"), (f32)(aCmp->km * M_4_PI));
     float scale = 1.0f / (aCmp->radius - aCmp->planetRadius);
     glUniform1f(m_program.getUniform("unScale"), scale);
     glUniform1f(m_program.getUniform("unScaleDepth"), aCmp->scaleDepth);
@@ -145,7 +145,7 @@ void GasGiantComponentRenderer::buildMesh() {
     m_numIndices = indices.size();
 
     std::vector<GasGiantVertex> vertices(positions.size());
-    for (int i = 0; i < positions.size(); i++) {
+    for (size_t i = 0; i < positions.size(); i++) {
         vertices[i].position = positions[i];
         vertices[i].texCoord = (positions[i].y + 1.0f) / 2.0f;
     }
